# api-documentation.yaml
openapi: 3.0.0
info:
  title: Party Picker API
  version: 1.0.0
...
security:
  - BearerAuth: []
  - BasicAuth: []
paths:
...
  /relations/{id}/messages:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        description: The ID of the relation
    get:
      summary: Get the messages of a conversation
      tags:
        - Relations
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            minimum: 0
            default: 0
          description: The page number to retrieve
        - in: query
          name: pageSize
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          description: The number of messages per page
      responses:
        "200":
          description: The messages were successfully retrieved
          headers:
            first:
              $ref: "#/components/headers/first"
            last:
              $ref: "#/components/headers/last"
            next:
              $ref: "#/components/headers/next"
            prev:
              $ref: "#/components/headers/prev"
          content:
            application/vnd.messages.v1+json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
        "404":
          description: The relation was not found
    post:
      summary: Send a message to a conversation
      tags:
        - Relations
      requestBody:
        description: The details of the new message
        required: true
        content:
          application/vnd.message.v1+json:
            schema:
              $ref: "#/components/schemas/Message"
      responses:
        "201":
          description: The message was successfully sent
          headers:
            Location:
              description: URI of the created resource
              schema:
                type: string
                format: uri
        "400":
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FormErrors"
        "404":
          description: The relation was not found
  /relations/{id}/messages/{messageId}:
    parameters:
      - in: path
        name: id
        required: true
        schema:
          type: integer
        description: The ID of the relation
      - in: path
        name: messageId
        required: true
        schema:
          type: integer
        description: The ID of the message
    get:
      summary: Get a message by ID
      tags:
        - Relations
      responses:
        "200":
          description: The message was successfully retrieved
          content:
            application/vnd.message.v1+json:
              schema:
                $ref: "#/components/schemas/Message"
        "404":
          description: The relation or message was not found
...
components:
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Message:
      type: object
      properties:
        self:
          type: string
          format: uri
          readOnly: true
          description: The URI to retrieve the full details of the message
        sender:
          type: string
          format: uri
          description: The URI of the sender
        receiver:
          type: string
          format: uri
          readOnly: true
          description: The URI of the receiver
        message:
          type: string
          description: The text of the message
        timestamp:
          type: string
          format: date-time
          readOnly: true
          description: The date of the message
        isRead:
          type: boolean
          readOnly: true
          description: Whether the message was read by the receiver
...
    FormErrors:
      type: array
      items:
        type: object
        properties:
          fieldName:
            type: string
            description: The name of the field that caused the error
          message:
            type: string
            description: The error message
...
  headers:
    first:
      schema:
        type: string
        format: uri
      description: The URI to the first page of the results
    last:
      schema:
        type: string
        format: uri
      description: The URI to the last page of the results
    next:
      schema:
        type: string
        format: uri
      description: The URI to the next page of the results, if there is one
    prev:
      schema:
        type: string
        format: uri
      description: The URI to the previous page of the results, if there is one
servers:
  - description: Party Picker API
    url: http://pawserver.it.itba.edu.ar/paw-2023b-04/api